jsim_n from Stellenbosch University
This was downloaded from the Stellenbosch University website
May 3, 2017

http://www0.sun.ac.za/ix/?q=tools_jsim

 JSIM_n source code, last modified by Mark Volkmann, 2014	124.15 KB

http://www0.sun.ac.za/ix/?q=system/files/attachment/jsim_n.tar.gz&file=1&type=node&id=19

Notes from the web page include the following:

JSIM and JSIM_n

JSIM and the noise-enabled JSIM_n are ubiquitous Josephson circuit simulators; faster than Spice with inherent support for the Josephson junction.

The source code with makefile is included as a tar.gz archive, and is easy to build on Linux and Mac platforms.

The original paper on JSIM is also included because it is very difficult to find. It should be referenced as: E. S. Fang and T. Van Duzer, "A Josephson integrated circuit simulator (JSIM) for superconductive electronics application," in Extended Abstracts of 1989 Intl. Superconductivity Electronics Conf. (ISEC '89), Tokyo, Japan: JSAP, 1989, pp. 407-410.

Attachment	Size
 Windows executables for JSIM and JSIM_n	122.02 KB
 JSIM user manual from Savoie University. (c) Pascal Febvre.	866.23 KB
 JSIM_n source code, last modified by Mark Volkmann, 2014	124.15 KB
 Original JSIM paper in Extended Abstracts of ISEC'89, by Emerson Fang	315.52 KB